+++ tests/src/Opulence/Authorization/Mocks/User.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2016 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+namespace Opulence\Tests\Authorization\Mocks;
+
+/**
+ * Defines a mock user
+ */
+class User
+{
+    /** @var int The user Id */
+    private $id = -1;
+
+    /**
+     * @param int $id The user Id
+     */
+    public function __construct(int $id)
+    {
+        $this->id = $id;
+    }
+
+    /**
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+}